+++ next.config.ts
@@ -1,10 +1,7 @@
 import type { NextConfig } from "next";
 
-// configs
-import i18n from "./i18n.config";
 
 const nextConfig: NextConfig = {
-  i18n,
   /* config options here */
 };
