+++ public/locales/en/common.json
@@ -8,6 +8,9 @@
         "confirmPassword": "Confirm password",
         "submit": "Sign Up",
         "successMessage": "Your account has been successfully created. A confirmation email has been sent to your address. Please check your inbox and click the confirmation link to activate your account.",
-        "errorMessage": "An error occurred while creating your account. Please try again later or contact support if the problem persists."
+        "errorMessage": "An error occurred while creating your account. Please try again later or contact support if the problem persists.",
+        "signin": "Sign In",
+        "alreadyHaveAccount": "Already have an account?",
+        "noAccount": "Don't have an account?"
     }
 }
