+++ public/locales/fr/common.json
@@ -8,6 +8,9 @@
         "confirmPassword": "Confirmer le mot de passe",
         "submit": "S'inscrire",
         "successMessage": "Votre compte a été créé avec succès. Un email de confirmation a été envoyé à votre adresse. Veuillez vérifier votre boîte de réception et cliquer sur le lien de confirmation pour activer votre compte.",
-        "errorMessage": "Une erreur s'est produite lors de la création de votre compte. Veuillez réessayer plus tard ou contacter le support si le problème persiste."
+        "errorMessage": "Une erreur s'est produite lors de la création de votre compte. Veuillez réessayer plus tard ou contacter le support si le problème persiste.",
+        "signin": "Se connecter",
+        "alreadyHaveAccount": "Vous avez déjà un compte ?",
+        "noAccount": "Vous n'avez pas de compte ?"
     }
 }
